function/requirements.txt
google-cloud-storage==2.10.0
google-cloud-firestore==2.11.0
google-cloud-vision==3.3.0
google-cloud-translate==3.11.0